Make some `StateScheduler` utils shared
