+++ denotational/core/src/test/kotlin/io/github/oxidefrp/core/CellOperatorsOperationalUnitTests.kt
@@ -1,6 +1,7 @@
 package io.github.oxidefrp.core
 
 import io.github.oxidefrp.core.shared.StateSchedulerLayer
+import io.github.oxidefrp.core.shared.pullEnter
 import io.github.oxidefrp.core.test_utils.tableSignal
 import kotlin.test.Test
 import kotlin.test.assertEquals
